backend_server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
Pillow==10.1.0
python-magic==0.4.27
fastapi-mail==1.4.1
jinja2==3.1.2
psycopg2-binary==2.9.9
email-validator==2.1.0.post1 